Refs #1172 and #1054 - fixed some caching and locking issues
